categories in redux
