+++ src/pages/CategoriesPage/CategoriesPage.jsx
@@ -1,9 +1,35 @@
+// node modules
+import { Suspense, useEffect } from 'react';
+
+// npm packages
+import { Outlet } from 'react-router-dom';
+import { useDispatch } from 'react-redux';
+
+// functions
+import { setCategories } from "../../redux/recipes/recipes.slices.js";
+import getRecipeCategories from "../../api/recipes/getRecipeCategories.js";
+
+// components
 import MainTitle from '../../components/MainTitle/MainTitle.jsx';
 import CategoriesNav from '../../components/CategoriesNav/CategoriesNav.jsx';
-import { Suspense } from 'react';
-import { Outlet } from 'react-router-dom';
 
 const CategoriesPage = () => {
+
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    const fetchCategories = async () => {
+      try {
+        const response = await getRecipeCategories();
+        dispatch(setCategories(response));
+        console.log("fetchCategories - response: ", response);
+      } catch (error) {
+        console.error("Error fetching categories:", error);
+      }
+    };
+    fetchCategories();
+  }, [])
+
   return (
     <>
       <MainTitle>Categories</MainTitle>
