+++ src/redux/recipes/recipes.initialState.js
@@ -1,5 +1,5 @@
 export const initialState = {
   category: 'Breakfast',
+  categories: [],
   // recipes: [],
-  // categories: []
 }
