+++ src/redux/recipes/recipes.selectors.js
@@ -1,5 +1,5 @@
 export const getCategory = (state) => state.recipes.category;
 
-// export const getRecipes = (state) => state.recipes.recipes;
+export const getCategories = (state) => state.recipes.categories
 
-// export const getCategories = (state) => state.recipes.categories
+// export const getRecipes = (state) => state.recipes.recipes;
